refactor(coding): add work shell adapter
