+++ src/loushang/coding/work_shell.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from loushang.work import CodingWorkShell
+
+__all__ = ["CodingWorkShell"]
