+++ tests/coding/test_work_shell_adapter.py
@@ -1,0 +1,8 @@
+from __future__ import annotations
+
+
+def test_coding_work_shell_adapter_exposes_coding_owned_entrypoint() -> None:
+    from loushang.coding.work_shell import CodingWorkShell
+    from loushang.work import CodingWorkShell as CompatCodingWorkShell
+
+    assert CodingWorkShell is CompatCodingWorkShell
